+++ crates/hcl-rs/src/format/mod.rs
@@ -39,6 +39,7 @@ mod impls;
 
 use self::escape::{CharEscape, ESCAPE};
 use crate::Result;
+use hcl_primitives::template::escape_markers;
 use std::io;
 
 mod private {
@@ -425,15 +426,17 @@ where
         self.write_bytes(s.as_bytes())
     }
 
-    /// Writes a quoted string to the writer. The quoted string will be escaped if `escape` is
-    /// true.
-    fn write_quoted_string(&mut self, s: &str, escape: bool) -> Result<()> {
+    /// Writes a quoted string to the writer.
+    fn write_quoted_string(&mut self, s: &str) -> Result<()> {
         self.write_bytes(b"\"")?;
-        if escape {
-            self.write_escaped_string(s)?;
-        } else {
-            self.write_string_fragment(s)?;
-        }
+        self.write_string_fragment(s)?;
+        self.write_bytes(b"\"")
+    }
+
+    /// Writes a quoted string to the writer after escaping it.
+    fn write_quoted_string_escaped(&mut self, s: &str) -> Result<()> {
+        self.write_bytes(b"\"")?;
+        self.write_escaped_string(s)?;
         self.write_bytes(b"\"")
     }
 
@@ -445,6 +448,7 @@ where
     /// Writes a string to the writer and escapes control characters and quotes that might be
     /// contained in it.
     fn write_escaped_string(&mut self, value: &str) -> Result<()> {
+        let value = escape_markers(value);
         let bytes = value.as_bytes();
 
         let mut start = 0;